+++ test/WebApiContribTests/Formatting/ReadWriteFormUrlEncodedFormatterTests.cs
@@ -80,5 +80,4 @@ namespace WebApiContribTests.Formatting
             }
         }
     }
-
 }
