+++ test/WebApiContribTests/Helpers/Contact.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace WebApiContribTests.Helpers
+{
+    public class Contact
+    {
+        public DateTime Birthday { get; set; }
+
+        public int Id { get; set; }
+    }
+}
