+++ test/WebApiContribTests/Helpers/IContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContribTests.Helpers
+{
+    public interface IContactRepository
+    {
+    }
+}
