+++ test/WebApiContribTests/Helpers/InMemoryContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContribTests.Helpers
+{
+    public class InMemoryContactRepository : IContactRepository
+    {
+    }
+}
