+++ test/WebApiContribTests/MessageHandlers/LoggingHandlerTests.cs
@@ -11,6 +11,7 @@ using WebApiContrib;
 using WebApiContrib.Formatting;
 using WebApiContrib.MessageHandlers;
 using WebApiContrib.Testing;
+using WebApiContribTests.Helpers;
 
 namespace WebApiContribTests.MessageHandlers
 {
@@ -56,13 +57,6 @@ namespace WebApiContribTests.MessageHandlers
             Assert.AreEqual(2, dummyRepository.LogMessageCount);
             Assert.IsTrue(dummyRepository.HasRequestMessageTypeBeenReceived, "No request message has been logged");
             Assert.IsTrue(dummyRepository.HasResponseMessageTypeBeenReceived, "No Response message has been received");
-        }
-
-        class Contact
-        {
-            public DateTime Birthday { get; set; }
-
-            public int Id { get; set; }
         }
     }
 }
