+++ test/WebApiContribTests/WebApiContribTests.csproj
@@ -81,6 +81,11 @@
   <ItemGroup>
     <Compile Include="Formatting\ReadWriteFormUrlEncodedFormatterTests.cs" />
     <Compile Include="Formatting\ServiceStackTextFormatterTests.cs" />
+    <Compile Include="Helpers\Contact.cs" />
+    <Compile Include="Helpers\ContactsController.cs" />
+    <Compile Include="Helpers\IContactRepository.cs" />
+    <Compile Include="Helpers\InMemoryContactRepository.cs" />
+    <Compile Include="Helpers\PrecannedMessageHandler.cs" />
     <Compile Include="IoC\DependencyInjectionTests.cs" />
     <Compile Include="MessageHandlers\EncodingHandlerTests.cs" />
     <Compile Include="MessageHandlers\HeadMessageHandlerTests.cs" />
